Time: 0 ms (100%), Space: 13.2 MB (64.04%) - LeetHub

+++ 3084-count-substrings-starting-and-ending-with-given-character/3084-count-substrings-starting-and-ending-with-given-character.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long countSubstrings(string s, char c) {
-        long cnt = 0;
+        long long cnt = 0;
         for(int i=0;i<s.size();i++){
             if(s[i]==c) cnt++;
         }
